refactor: move plural-maybe.ts to common utils
